Adding work for LRO pagers
